Simplify userop signing flow
